refactor(backend): split contacts crud.py into bulk_ops + enrichment (728 -> 410 lines)
